Horrible hack to address short-term linux problem
For some strange reason, realpath is hanging if a file does not exist.
(It should return a "file does not exist code").  The hack is to attempt
to create the file if it does not exist and *then* asking for the absolute
path.  Bad hack, but it seems to work.