Action "state-leaving" behavior refactored
When an agent leaves a state, the action allows for a boolean "reset".
This change allows a more general change.  Specifically, upon exiting
the parent state, the action class can do two things: one is dependent
on the reset flag (formerly called "leaveAction", now called "resetAction")
and another action which will be invoked unconditionally (the new
"leaveAction".  It is guaranteed that the resetAction will be invoked
*before* the the leaveAction (if invoked at all).