Implementation of GAS stress model as a plugin (#23)
* Add initial gas stress plug-in

This represents the initial gas action -- it applies stress to agents.

1) Kind of clunky
2) Can't be configured.

* Add stress demo

This is a variation of the periodic demo but with stress added.

* Add stress plugin to linux build

Add the appropriate cmake infrastructure.

* Correct linux compatibility

Account for file case and use proper linux-compatible directory slashes.

* Cosmetics clean up

Scrubbed code documentation.  Made sure it was of appropriate format and
confirmed that the comments actually referred to this code (as opposed
to the source it was copied from.)

* New stress action architecture

Whole new system for accumulating stress.

1) BaseStressAction is responsible for registering agents with the manager,
   handling exit behavior, defining the stress vector.
   a) Sub-class it to apply different stress functions to agents.
2) Stress functions define how stress is defined.  This includes a single
   kind of stress function: timed stress function.  Stress simply accumulates
   at a linear rate up to 100% stress.

Removed old action, re-wired manager, modified sample scenarios.

WIP: In release mode, it arbitrarily crashes.  I need to figure out why
     that is.

* Modify stress enumeration access to match C++ access

I had used C# style accessors (e.g., EnumName::EnuMValue).  Visual studio
is content with this, but G++ gets cranky.

* Correct include path to be linux compatible

I had backslashes that needed to be forward.

* Improve attribute set parsing warning messages

If a distribution is expected to be read and it is not found, the warning
message now includes the prefix.

* Added missing lock to manager

Because the manager gets executed in parallel (agents can enter and exit)
it is important that the underlying hash map be protected.  Introduced a
readers/writer lock to allow multiple readers and single writer
(eventually).

* Correct cool down function

The cool down function now properly performs cool down based on time step.
The cool down period can also be specified in XML.

* Added sample of the cooldown functionality

It shows three agents:
  1) One agent continues accumulating stress
  2) One cools down in 5 seconds
  3) One cools down in 10 seconds.

* Rename StressTasks to StressTask

(Changed plural to singular because there is only a single task.)

* Add dll export to more of the classes

* Update doxygen material

The stress functionality is now documented and the validity has been
confirmed.

* Stress ending demo

This demo shows the exit_behavior of the stress action, showing different
behaviors (continue vs cool down)  and different cool down rates.

* Correct doxygen type

* Update gitignore for my local build

* Initial implementation of distance stress function

Largely just a place holder; actual functionality still missing.

* Updated release notes

* Manager removes stress functions when cool down is completed

The stress function signals that it is done and will not be evaluated
again.

* Cleaned up zombie line of code.

* Add distance stress function

An action will initiate stress based on proximity to a region.  This
includes the DistanceStressAction, DistanceStressFunction, and a demo.

* Doxygen correction

* Updated release notes

* Update run all demos scripts

Add the stress demos to the various scripts.