Add agent context sub-classes
This is the first step in recreating the functionality to have contexts
for arbitrary element implementations.

This adds a visual element database (somewhat mirroring the Element
database in the simulator.)  In this case, it pulls a context for a type
of simulation element.  This does *not* yet have plug-in functionality
but it does let me define custom contexts for pedvo and orca and have it
pull up correctly.

This is also a step towards having hybrid agent populations; the context
is selected at run time based on the agent type (rather than at compile
time as part of the system.)