Velocity Component setPrefVelocity is now a const method
There is an issue; the road map velocity component *may* end up updating
the paths based on the state at the time of the query.  This requires the
path to be mutable.

TODO: Perform this work in a task, so that the operations that should *be*
const don't require mutable members.