Gcf (#47)
Introduces the Generalized Centrifugal Force (GCF) model.

This is a force-based model where the agents are represented as ellipses instead of circles. The ellipse shape is a function of travel speed.  This model exposed a number of issues, some corrected in this PR, and some left as exercises for the future.

This PR includes a modification of the orientation computation and correction of nav mesh placement.