Initial hijack example
Provides and example where an agent is pulled of its roadmap. It is swept
along by a larger crowd and ends up in a cul-de-sac which he can't get out
of. The reason for this is because he doesn't replan and only knows the
last position he was in when he *could* see his path. If there is no
direct path to that position, the agent gets caught in a local minimum.