Roadmap path following triggers replan upon path loss
The previous method had a stopgap measure to make path following *slightly*
robust to agents being pushed off their path. However, pushed sufficiently
far, an agent would lose the ability to return to the path.

This fix makes it so the agent recognizes when this has happened and
triggers a replan. If the agent is pushed to a point where *no* road map
waypoints are visible, an exception is thrown.