Bugfix - Agent can't find path while colliding with an obstacle.
The process of finding a path on a roadmap for an agent consists of finding
"visible" links from the agent position to a road map vertex. The
visibility query w.r.t. a *single* obstacle essentially tests for
intersection between the obstacle's line segment and a capsule formed by
the desired link (agent position to vertex position) convolved with a
disk of the agent's radius. For a given vertex, if any obstacle intersects
the vertex is not "visible". If the agent is intersecting an obstacle,
no vertex will be considered visible.

The fix:
  1. Add new query to SpatialQuery; linkIsTraversible(). It is closely
     related to the queryVsibility() method with one extra detail; it
     handles the case where the agent is intersecting an obstacle.
  2. Extend the API through the two implementations of the interface
     (the nav mesh implementation is *not* implemented at all.) The kd-tree
     implementaiton is, essentially, a copy of the visibility query with
     the last test modified.
  3. Change the calls in the roadmap from visibility to traversibility.
  4. Add some stray clarifying comments.
  5. Remove asserts and replace them with more appropriate exceptions.